+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [INTERNAL] npm audit fix
 
+### Changed
+- Update textile-js to GehDoc/textile-js@8a6755291e1e8eb9e73620cbe26c7c1c19fc3343, to get new applyHooks function, and some fixes for `<LI>` and `<HR>`
+
 
 ## [1.2.0] - 20200323
 ### Changed